+ Added Patron Capes. % Changed Power Sub-Station recipe, harder in GTNH. Closes #194. Also closes https://github.com/GTNewHorizons/NewHorizons/issues/2517. % Tweaked Industrial Sifter & Wash Plant items/operation. % Greatly improved the Iron Blast Furnace for GT 5.09. It is now double the speed of a Bronze/Bricked Blast Furnace.

+++ src/Java/gtPlusPlus/core/util/Utils.java
@@ -566,7 +566,7 @@ public class Utils {
 
 	}
 
-	public static int calculateVoltageTier(final int Voltage) {
+	public static int calculateVoltageTier(final long Voltage) {
 		int V;
 		if (Voltage == 8) {
 			V = 0;
